app/requirements.txt
fastapi
uvicorn[standard]
websockets
azure-ai-voicelive[aiohttp]
azure-identity
python-dotenv
